refactor(api-scheduler): move common context wiring into base package
The scheduler context initialization (model registration, permissions,
GraphQL factory, features) was duplicated in both api-scheduler-aws and
api-scheduler-server. Extracted into api-scheduler/src/context.ts so
implementation packages only register their SchedulerService.
