+++ packages/api-scheduler-aws/__tests__/__mocks/handler/plugins.ts
@@ -17,7 +17,7 @@ import apiKeyAuthorization from "@webiny/api-core/legacy/security/plugins/apiKey
 import { createApiCore } from "@webiny/api-core";
 import type { IdentityData } from "@webiny/api-core/features/security/IdentityContext/index.js";
 import type { ApiCoreStorageOperations } from "@webiny/api-core/types/core.js";
-import { createSchedulerContext } from "~/context.js";
+import { createAwsSchedulerContext } from "~/context.js";
 
 export interface CreateHandlerCoreParams {
     getScheduleClient: (config?: SchedulerClientConfig) => Pick<SchedulerClient, "send">;
@@ -64,7 +64,7 @@ export const createHandlerCore = (params: CreateHandlerCoreParams) => {
             createHeadlessCmsGraphQL(),
             plugins,
             graphQLHandlerPlugins(),
-            createSchedulerContext({
+            createAwsSchedulerContext({
                 getClient: config => {
                     return params.getScheduleClient(config);
                 }
