+++ packages/api-scheduler-aws/src/createScheduler.ts
@@ -1,12 +1,14 @@
 import type { Plugin } from "@webiny/plugins/types.js";
-import { SchedulerClient, SchedulerClientConfig } from "@webiny/aws-sdk/client-scheduler/index.js";
-import { createSchedulerContext } from "~/context.js";
+import type { SchedulerClientConfig } from "@webiny/aws-sdk/client-scheduler/index.js";
+import { createAwsSchedulerContext } from "~/context.js";
 import { createScheduledActionEventHandler } from "~/createEventHandler.js";
 
 export interface ICreateSchedulerParams {
-    getClient(config?: SchedulerClientConfig): Pick<SchedulerClient, "send">;
+    getClient(
+        config?: SchedulerClientConfig
+    ): Pick<import("@webiny/aws-sdk/client-scheduler/index.js").SchedulerClient, "send">;
 }
 
 export const createScheduler = (params: ICreateSchedulerParams): Plugin[] => {
-    return [createScheduledActionEventHandler(), ...createSchedulerContext(params)];
+    return [createScheduledActionEventHandler(), ...createAwsSchedulerContext(params)];
 };
