+++ packages/api-scheduler-server/src/createScheduler.ts
@@ -1,6 +1,6 @@
 import type { Plugin } from "@webiny/plugins/types.js";
-import { createSchedulerContext } from "~/context.js";
+import { createServerSchedulerContext } from "~/context.js";
 
 export const createScheduler = (): Plugin[] => {
-    return [...createSchedulerContext()];
+    return [...createServerSchedulerContext()];
 };
