+++ packages/api-scheduler/package.json
@@ -22,6 +22,7 @@
     "@webiny/api-core": "0.0.0",
     "@webiny/api-headless-cms": "0.0.0",
     "@webiny/feature": "0.0.0",
+    "@webiny/handler": "0.0.0",
     "@webiny/handler-graphql": "0.0.0",
     "@webiny/utils": "0.0.0"
   },
